Fix a few tests
